feat(airdrops): Cast claimed_count to int
Casts the `claimed_count` to an integer in the airdrops API response. This ensures consistent typing for the count, preventing potential issues downstream.

+++ api/airdrops.ts
@@ -116,7 +116,7 @@ async function handleGet(req: VercelRequest, res: VercelResponse) {
         const { rows } = await sql`
             SELECT 
                 a.*,
-                (SELECT COUNT(*) FROM claims c WHERE c.airdrop_id = a.id) as claimed_count
+                (SELECT COUNT(*)::int FROM claims c WHERE c.airdrop_id = a.id) as claimed_count
             FROM airdrops a 
             ORDER BY a.created_at DESC;
         `;
